papelera funcionando, faltan detalles
